Try reading PdxInstance values as arrays first
When we read them as non-arrays first, we did not reliably get an
exception. This would result in corrupted data ended up in the 
CacheablePtr. It appeared that the size of the array would be
interpreted as the typeId of the CacheablePtr in several instances.

So we are trying out reading as CacheableObjectArrayPtr first because as
of this time we have not yet seen getField create a corrupt
CacheableObjectArrayPtr. Instead, we reliably receive exceptions.

[Finishes #81718204]